GitButler Integration Commit
This is an integration commit for the virtual branches that GitButler is tracking.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily.

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - lab (refs/gitbutler/lab)
   branch head: 2856498f0520e83a35c48300eb90ec3da3b4c08c
   - lc3os_v2.asm
   - starter_code.asm
   - readme.md

Your previous branch was: refs/heads/main

The sha for that commit was: 1c822edff3192746ae51ec5026996373cabecda5

For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch